pyproject.toml
[project]
name = "luckypot"
version = "0.1.0"
description = "LuckyPot is a bot that implments a lottery system ontop of StackCoin"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hikari>=2.3.5",
    "hikari-lightbulb>=3.1.1",
    "python-dotenv>=1.1.1",
]
